chore: Update video page to display file extension and shorten file name
This commit updates the video page to display the file extension of the video being fetched. It also adds a function to shorten the file name if it exceeds 20 characters. This improves the user experience by providing more information about the video files and ensuring that long file names are properly displayed.